+++ Geotiff/TileOrStripResult.cs
@@ -4,7 +4,7 @@ namespace Geotiff;
 
 internal class TileOrStripResult
 {
-    public int x { get; set; }
-    public int y { get; set; }
+    public ulong x { get; set; }
+    public ulong y { get; set; }
     public byte[] data { get; set; }
 }
